+++ lib/stripped-build.js
@@ -27,6 +27,7 @@ module.exports = function(packageName, tree, _options) {
     filterImports({
       'ember-data/-private/debug': [
         'assert',
+        'assertPolymorphicType',
         'debug',
         'deprecate',
         'info',
