+++ tests/unit/utils-test.js
@@ -6,7 +6,7 @@ import {module, test} from 'qunit';
 import DS from 'ember-data';
 import Model from 'ember-data/model';
 
-import { assertPolymorphicType } from "ember-data/-private/utils";
+import { assertPolymorphicType } from "ember-data/-private/debug";
 import { modelHasAttributeOrRelationshipNamedType } from "ember-data/-private/utils";
 
 var env, User, Message, Post, Person, Video, Medium;
